# CRUDClient/Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CRUDAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Pessoas": {
      "get": {
        "tags": [
          "Pessoas"
        ],
        "operationId": "GetPessoas",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pessoa"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pessoa"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pessoa"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pessoas"
        ],
        "operationId": "PostPessoa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pessoas/{id}": {
      "get": {
        "tags": [
          "Pessoas"
        ],
        "operationId": "GetPessoa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pessoas"
        ],
        "operationId": "PutPessoa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Pessoa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pessoas"
        ],
        "operationId": "DeletePessoa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Pessoa": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "sobrenome": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}